Move SigningConfig and signing::traits to the top of the sign module as they relate to signing in general, and rename the signing sub-module as it is specific to signature generation.
