+++ src/sign/signatures/mod.rs
@@ -1,0 +1,3 @@
+//! Signature generation.
+pub mod rrsigs;
+pub mod strategy;
