+++ src/sign/signatures/strategy.rs
