--- src/sign/signing/mod.rs
@@ -1,6 +1,0 @@
-pub mod config;
-pub mod rrsigs;
-pub mod strategy;
-pub mod traits;
-
-pub use self::config::SigningConfig;
